fix: dataTables and select2 styles

+++ resources/views/layouts/dropzone_js.blade.php
@@ -5,23 +5,29 @@
     Dropzone.autoDiscover = false;
     $(document).ready(function() {
         $('.dropzone').each(function () {
-            var _el = $(this);
-            var myDropzone = new Dropzone(`#${$(this).attr('id')}`, {
+            const _el = $(this);
+            const isMultiple = typeof _el.data('multiple') !== 'undefined' ? true : false;
+            let myDropzone = new Dropzone(`#${$(this).attr('id')}`, {
                 accept: function(file, done) {
                     console.log("uploaded");
                     done();
                 },
                 acceptedFiles: typeof _el.data('accept') !== 'undefined' ? _el.data('accept') : '*',
                 init: function() {
-                    this.on("addedfile", function() {
-                        if (this.files[1]!=null){
-                            this.removeFile(this.files[0]);
-                            $(document).find('.file-path').html('');
-                        }
-                    });
+                    if (!isMultiple) {
+                        this.on("addedfile", function() {
+                            if (this.files[1]!=null){
+                                this.removeFile(this.files[0]);
+                                $(document).find('.file-path').html('');
+                            }
+                        });
+                    }
+                    $(document).find('.file-path').html('');
                 },
                 url: "{{ route('drop_zone_file_upload') }}",
                 timeout: 0,
+                parallelUploads: 10,
+                uploadMultiple: isMultiple,
                 addRemoveLinks: false,
                 dictDefaultMessage: `<div><h6>Drag and drop to upload</h6><h6>or <a href="javascript:void(0)">browse</a> to choose a file</h6></div>`,
                 sending: function(file, xhr, formData) {
@@ -32,10 +38,17 @@
                     }
                 },
                 success: function(file, response) {
-                    console.log(response);
-                    $(document).find('.file-path').html(`
-                    <input name="${_el.data('name')}" value="${response.data.imagePath}">
-                    `);
+                    if (typeof response.data.imagePath === 'string') {
+                        $(document).find('.file-path').html(`
+                        <input name="${_el.data('name')}" value="${response.data.imagePath}">
+                        `);
+                    } else if (typeof response.data.imagePath === 'object') {
+                        let content = '';
+                        response.data.imagePath.forEach((imagePath) => {
+                            content += `<input name="${_el.data('name')}[]" value="${imagePath}">`
+                        })
+                        $(document).find('.file-path').html(content);
+                    }
                 },
                 error: function(file, errormessage, response) {
                     console.log(response);
